+++ changelog.md
@@ -6,6 +6,7 @@
 - x, y, cx, cy attribute defaults for circles, ellipses & rectangles
 - the optional 2nd & 3rd element for a rotation transformation, which were previously ignored, are now taken into
 account
+- skewY transformations are now expressed through a matrix transformation, fixing unexpected Imagick behaviour
 
 ## [1.2.0](https://github.com/bartdecorte/imagick-svg/compare/1.1.0...1.2.0) - 2022-03-15